tested group merge with all the cythonized components

+++ spark_fof_c.pyx
@@ -78,7 +78,7 @@ def partition_particles_cython(particles, domain_containers, tau):
     p_arr = np.fromiter(particles, pdt)
     nparts = len(p_arr)
     
-    trans = np.array([[-tau, 0, 0], [0,-tau, 0], [0, 0, -tau], [-tau, -tau, 0], [0, -tau, -tau], [-tau,-tau,-tau]])
+    trans = np.array([[-tau, 0, 0], [0,-tau, 0], [0, 0, -tau], [-tau, -tau, 0], [0, -tau, -tau], [-tau,-tau,-tau]], dtype=np.float32)
 
     dom_mins = np.array([-1.,-1.,-1.])
     dom_maxs = np.array([1., 1., 1.])
